feat: moved hd2017 to its own module
