+++ pysm/models/__init__.py
@@ -1,6 +1,7 @@
 # flake8: noqa
 
-from .dust import ModifiedBlackBody, DecorrelatedModifiedBlackBody, HensleyDraine2017
+from .dust import ModifiedBlackBody, DecorrelatedModifiedBlackBody
+from .hd2017 import HensleyDraine2017
 from .power_law import PowerLaw, CurvedPowerLaw
 from .template import Model, read_map, apply_smoothing_and_coord_transform
 from .spdust import SpDust, SpDustPol
